caplinked_youtube_automation/requirements.txt
requests
beautifulsoup4
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
